# backend/openapi/mlms.openapi.yaml
openapi: 3.0.3
info:
  title: MLMS API
  version: 1.0.0
  description: |
    Medical Laboratory Management System API.

    Supports the complete laboratory workflow:
    - Patient registration
    - Order creation
    - Specimen collection
    - Result entry
    - Report validation and publication
    - Patient portal access

    All protected endpoints require a Bearer JWT token obtained via `POST /auth/login`.
servers:
  - url: http://localhost:3000/api
    description: Local development server
...
security:
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      description: Authenticate a user and receive a JWT access token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
            example:
              email: physician@mlms.local
              password: Doctor1234!
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseDto'
              example:
                accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example
                user:
                  id: clxyz001
                  email: physician@mlms.local
                  role: PHYSICIAN
                  language: FR
                  isActive: true
                  createdAt: '2026-03-01T08:00:00.000Z'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new internal user
      description: Create a new user account. Requires ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserDto'
            example:
              email: technician2@mlms.local
              password: Tech1234!
              role: TECHNICIAN
              language: FR
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current authenticated user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurrentUserDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ── Patients ───────────────────────────────────────────────────────────────
  /patients:
    post:
      tags: [Patients]
      summary: Register a new patient
      description: Requires RECEPTION or ADMIN role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientRequestDto'
            example:
              firstName: Amina
              lastName: El Idrissi
              birthDate: '1992-05-12T00:00:00.000Z'
              gender: F
              phone: '+212600000001'
              email: amina@example.com
              address: Casablanca, Maroc
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [Patients]
      summary: List patients
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - in: query
          name: query
          schema:
            type: string
          description: Search by name or email
      responses:
        '200':
          description: Paginated patient list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PatientDto'
                  meta:
                    $ref: '#/components/schemas/ApiListMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /patients/{id}:
    get:
      tags: [Patients]
      summary: Get patient by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Patient found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

    patch:
      tags: [Patients]
      summary: Update patient
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientRequestDto'
      responses:
        '200':
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ── Orders ─────────────────────────────────────────────────────────────────
  /orders:
    post:
      tags: [Orders]
      summary: Create a new order
      description: Requires RECEPTION or ADMIN role. Order starts with status PENDING.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequestDto'
            example:
              patientId: clxyz001
              priority: ROUTINE
              tests:
                - testCode: GLU
                  testNameFr: Glycémie
                  testNameAr: سكر الدم
                  priority: ROUTINE
                - testCode: HGB
                  testNameFr: Hémoglobine
                  testNameAr: الهيموغلوبين
                  priority: ROUTINE
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          description: Patient not found
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'

    get:
      tags: [Orders]
      summary: List orders
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - in: query
          name: patientId
          schema:
            type: string
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - in: query
          name: priority
          schema:
            $ref: '#/components/schemas/OrderPriority'
      responses:
        '200':
          description: Paginated order list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrderDto'
                  meta:
                    $ref: '#/components/schemas/ApiListMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get order by ID
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /orders/{id}/status:
    patch:
      tags: [Orders]
      summary: Update order status
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrderStatusRequestDto'
            example:
              status: COLLECTED
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /orders/{id}/cancel:
    patch:
      tags: [Orders]
      summary: Cancel an order
      description: Cancels an order in PENDING, COLLECTED, or ANALYZED status. Requires RECEPTION or ADMIN role.
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Order cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /orders/{id}/results:
    get:
      tags: [Results]
      summary: Get all results for an order
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Results for order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResultDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /orders/{id}/specimens:
    get:
      tags: [Specimens]
      summary: Get all specimens for an order
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Specimens for order
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpecimenDto'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  # ── Specimens ──────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Parameters ──────────────────────────────────────────────────────────────
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: Resource ID (cuid)
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PageSizeParam:
      in: query
      name: pageSize
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page

  # ── Responses ───────────────────────────────────────────────────────────────
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          example:
            error:
              code: FORBIDDEN
              message: Access denied
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
    ConflictError:
      description: Conflict – resource already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          example:
            error:
              code: CONFLICT
              message: Resource already exists
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorEnvelope'
          example:
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                - field: firstName
                  message: firstName is required

  # ── Schemas ─────────────────────────────────────────────────────────────────
  schemas:
...
    Language:
      type: string
      enum: [FR, AR]
    UserRole:
      type: string
      enum: [RECEPTION, TECHNICIAN, PHYSICIAN, ADMIN]
    Gender:
      type: string
      enum: [M, F, O]
    OrderPriority:
      type: string
      enum: [ROUTINE, URGENT, STAT]
    OrderStatus:
      type: string
      enum: [PENDING, COLLECTED, ANALYZED, VALIDATED, PUBLISHED, CANCELLED]
    SpecimenType:
      type: string
      enum: [BLOOD, URINE, STOOL, TISSUE]
    SpecimenStatus:
      type: string
      enum: [COLLECTED, RECEIVED, PROCESSED, DISPOSED, REJECTED]
    ResultFlag:
      type: string
      enum: [N, H, L, HH, LL, CRITICAL]
...
    ApiFieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: firstName
        message:
          type: string
          example: firstName is required
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            $ref: '#/components/schemas/ApiFieldError'
    ApiErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'

    # ── Pagination ─────────────────────────────────────────────────────────
    ApiListMeta:
      type: object
      required: [page, pageSize, total, totalPages]
      properties:
        page:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 20
        total:
          type: integer
          example: 120
        totalPages:
          type: integer
          example: 6

    # ── Auth ───────────────────────────────────────────────────────────────
    LoginDto:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: physician@mlms.local
        password:
          type: string
          format: password
          example: secret123
    RegisterUserDto:
      type: object
      required: [email, password, role]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        role:
          $ref: '#/components/schemas/UserRole'
        language:
          $ref: '#/components/schemas/Language'
    CurrentUserDto:
      type: object
      required: [id, email, role, language, isActive, createdAt]
      properties:
        id:
          type: string
          example: clxyz001
        email:
          type: string
          format: email
          example: physician@mlms.local
        role:
          $ref: '#/components/schemas/UserRole'
        language:
          $ref: '#/components/schemas/Language'
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: '2026-03-01T08:00:00.000Z'
    LoginResponseDto:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example
        user:
          $ref: '#/components/schemas/CurrentUserDto'

    # ── Patient ────────────────────────────────────────────────────────────
    PatientDto:
      type: object
      required: [id, firstName, lastName, birthDate, gender, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: clxyz001
        firstName:
          type: string
          example: Amina
        lastName:
          type: string
          example: El Idrissi
        birthDate:
          type: string
          format: date-time
          example: '1992-05-12T00:00:00.000Z'
        gender:
          $ref: '#/components/schemas/Gender'
        phone:
          type: string
          nullable: true
          example: '+212600000001'
        email:
          type: string
          format: email
          nullable: true
          example: amina@example.com
        address:
          type: string
          nullable: true
          example: Casablanca, Maroc
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePatientRequestDto:
      type: object
      required: [firstName, lastName, birthDate, gender]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          $ref: '#/components/schemas/Gender'
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        address:
          type: string
          nullable: true
    UpdatePatientRequestDto:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        birthDate:
          type: string
          format: date-time
        gender:
          $ref: '#/components/schemas/Gender'
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        address:
          type: string
          nullable: true

    # ── Order ──────────────────────────────────────────────────────────────
    TestOrderItemDto:
      type: object
      required: [id, orderId, testCode, testNameFr, testNameAr, priority]
      properties:
        id:
          type: string
        orderId:
          type: string
        testCode:
          type: string
          example: GLU
        testNameFr:
          type: string
          example: "Glycémie"
        testNameAr:
          type: string
          example: "سكر الدم"
        priority:
          type: string
          enum: [ROUTINE, URGENT]
        notes:
          type: string
          nullable: true
    OrderDto:
      type: object
      required: [id, patientId, status, priority, createdAt, updatedAt, tests]
      properties:
        id:
          type: string
        patientId:
          type: string
        status:
          $ref: '#/components/schemas/OrderStatus'
        priority:
          $ref: '#/components/schemas/OrderPriority'
        createdBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        tests:
          type: array
          items:
            $ref: '#/components/schemas/TestOrderItemDto'
    CreateOrderRequestDto:
      type: object
      required: [patientId, priority, tests]
      properties:
        patientId:
          type: string
        priority:
          $ref: '#/components/schemas/OrderPriority'
        tests:
          type: array
          minItems: 1
          items:
            type: object
            required: [testCode, testNameFr, testNameAr, priority]
            properties:
              testCode:
                type: string
                example: GLU
              testNameFr:
                type: string
                example: "Glycémie"
              testNameAr:
                type: string
                example: "سكر الدم"
              priority:
                type: string
                enum: [ROUTINE, URGENT]
              notes:
                type: string
                nullable: true
    UpdateOrderStatusRequestDto:
      type: object
      required: [status]
      properties:
        status:
          $ref: '#/components/schemas/OrderStatus'

    # ── Specimen ───────────────────────────────────────────────────────────
    SpecimenDto:
      type: object
      required: [id, orderId, barcode, type, status, collectionTime, createdAt, updatedAt]
      properties:
        id:
          type: string
        orderId:
          type: string
        barcode:
          type: string
          example: LAB-2026-00001
        type:
          $ref: '#/components/schemas/SpecimenType'
        status:
          $ref: '#/components/schemas/SpecimenStatus'
        collectionTime:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ResultDto:
      type: object
      required: [id, specimenId, testCode, testNameFr, testNameAr, value, flag, measuredAt, createdAt, updatedAt]
      properties:
        id:
          type: string
        specimenId:
          type: string
        testCode:
          type: string
          example: GLU
        testNameFr:
          type: string
          example: "Glycémie"
        testNameAr:
          type: string
          example: "سكر الدم"
        value:
          type: string
          example: '7.2'
        unit:
          type: string
          nullable: true
          example: mmol/L
        referenceLow:
          type: number
          nullable: true
          example: 3.9
        referenceHigh:
          type: number
          nullable: true
          example: 6.1
        flag:
          $ref: '#/components/schemas/ResultFlag'
        measuredAt:
          type: string
          format: date-time
        measuredBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...